langchain/requirements.txt
langchain-community
langchain
langchain-openai
openai
PyPDF2
faiss-cpu
tiktoken
python-dotenv
flask
flask_cors